refactor(infra): repository.rs をモジュール分割・load_or_create 削除・TC-UT-142 assertion 修正
ペガサス・ペテルギウス指摘 4 点を反映（PR #124 rev.2）:

1. repository.rs（617行）を repository/mod.rs + repository/tests.rs に分離
   - 本体（本番コード）と UT を物理的に分割し見通しを改善
2. 旧 load_or_create メソッドを削除（DRY 違反・デッドコード）
   - load_or_create_plaintext が同責務を引継ぎ済み
   - TC-U-REPO-03（load_or_create を呼ぶテスト）も同時に削除
3. TC-UT-142 アサーションを型限定に修正:
   - 誤: assert!(result.is_err(), ...)
   - 正: assert!(matches!(result, Err(PersistenceError::InvalidPermission { .. })), ...)
4. TC-UT-142 NOTE コメントから Io(_) 誤記を除去し、verify_dir が先行する理由を明記
