+++ crates/shikomi-infra/src/persistence/repository/mod.rs
@@ -78,9 +78,8 @@ impl SqliteVaultRepository {
 
     /// vault を読み込む。`vault.db` が存在しない場合は空の plaintext vault を**永続化して**返す。
     ///
-    /// daemon コンポジションルートから呼び出す（REQ-DAEMON-028）。`load_or_create` との
-    /// 違いは、NotFound 時に空 vault を `save` で永続化し、`shikomi_daemon::init` ターゲットへ
-    /// 2 行のログを出力する点にある。
+    /// daemon コンポジションルートから呼び出す（REQ-DAEMON-028）。`NotFound` 時に空 vault を
+    /// `save` で永続化し、`shikomi_daemon::init` ターゲットへ 2 行のログを出力する。
     ///
     /// - vault.db が存在する → `load()` の結果をそのまま返す（ログ出力なし）
     /// - vault.db が存在しない → 空の plaintext vault を生成・保存してから返す
@@ -116,37 +115,6 @@ impl SqliteVaultRepository {
                     "hint: to enable encryption, run `shikomi vault encrypt` after the daemon has started"
                 );
                 Ok(vault)
-            }
-            Err(e) => Err(e),
-        }
-    }
-
-    /// vault を読み込む。`vault.db` が存在しない場合は空の plaintext vault を返す。
-    ///
-    /// 初回インストール / CI 環境など `vault.db` が未作成の状態で daemon が起動する場合に
-    /// IPC を受け付けられるよう、空の vault で起動する。NotFound 以外のエラーは伝播する。
-    ///
-    /// # Errors
-    ///
-    /// - `vault.db` が存在しない以外の IO エラー / 破損データ / ロック取得失敗:
-    ///   `PersistenceError`
-    pub fn load_or_create(&self) -> Result<Vault, PersistenceError> {
-        match self.load() {
-            Ok(v) => Ok(v),
-            Err(PersistenceError::Io { ref source, .. })
-                if source.kind() == std::io::ErrorKind::NotFound =>
-            {
-                tracing::info!(
-                    target: "shikomi_infra::persistence",
-                    "vault.db not found — starting with empty plaintext vault (new installation)"
-                );
-                Ok(Vault::new(
-                    VaultHeader::new_plaintext(
-                        VaultVersion::CURRENT,
-                        time::OffsetDateTime::now_utc(),
-                    )
-                    .expect("CURRENT version is always valid"),
-                ))
             }
             Err(e) => Err(e),
         }
@@ -431,187 +399,5 @@ impl SqliteVaultRepository {
     }
 }
 
-// ---------------------------------------------------------------------------
-// ユニットテスト
-// ---------------------------------------------------------------------------
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    // --- TC-U-REPO-01: prepare_dir — 存在しないディレクトリを作成する ---
-
-    #[test]
-    fn tc_u_repo_01_prepare_dir_creates_directory() {
-        let base = TempDir::new().unwrap();
-        let vault_dir = base.path().join("vault_new");
-
-        // 事前条件: ディレクトリが存在しない
-        assert!(!vault_dir.exists(), "事前条件: ディレクトリが未作成");
-
-        let repo = SqliteVaultRepository::from_directory(&vault_dir).unwrap();
-        repo.prepare_dir().expect("prepare_dir は成功すべき");
-
-        assert!(
-            vault_dir.is_dir(),
-            "prepare_dir 後: ディレクトリが作成されるべき"
-        );
-    }
-
-    // --- TC-U-REPO-02: prepare_dir — 既存ディレクトリに冪等 ---
-
-    #[test]
-    fn tc_u_repo_02_prepare_dir_is_idempotent_on_existing_dir() {
-        let dir = TempDir::new().unwrap();
-        let repo = SqliteVaultRepository::from_directory(dir.path()).unwrap();
-
-        // 2 回呼び出しても失敗しない
-        repo.prepare_dir()
-            .expect("1 回目: prepare_dir は成功すべき");
-        repo.prepare_dir()
-            .expect("2 回目: prepare_dir は冪等であるべき");
-    }
-
-    // --- TC-U-REPO-03: load_or_create — vault.db 未存在時に空 plaintext vault を返す ---
-
-    #[test]
-    fn tc_u_repo_03_load_or_create_returns_empty_vault_when_not_found() {
-        let dir = TempDir::new().unwrap();
-        let repo = SqliteVaultRepository::from_directory(dir.path()).unwrap();
-        // ディレクトリのパーミッションを OS 規定に設定してから load_or_create を呼ぶ
-        repo.prepare_dir().expect("prepare_dir は成功すべき");
-
-        // vault.db が存在しない → NotFound を捕捉して空 plaintext vault を返すべき
-        let vault = repo
-            .load_or_create()
-            .expect("load_or_create は空 vault を返すべき");
-
-        assert!(vault.records().is_empty(), "初期 vault はレコード 0 件");
-        assert!(
-            matches!(
-                vault.protection_mode(),
-                shikomi_core::ProtectionMode::Plaintext
-            ),
-            "初期 vault は plaintext モードであるべき"
-        );
-    }
-
-    // ---------------------------------------------------------------------------
-    // TC-UT-140〜142: SqliteVaultRepository::load_or_create_plaintext
-    // 設計書: docs/features/daemon-ipc/test-design/unit.md §2.18
-    // REQ-DAEMON-028 / Issue #80
-    // ---------------------------------------------------------------------------
-
-    // --- TC-UT-140: vault.db 存在時 — 既存 vault をロードして返す（ログ出力なし） ---
-
-    #[test]
-    #[tracing_test::traced_test]
-    fn tc_ut_140_load_or_create_plaintext_returns_existing_vault_without_log() {
-        // REQ-DAEMON-028 / Issue #80
-        let dir = TempDir::new().unwrap();
-        let repo = SqliteVaultRepository::from_directory(dir.path()).unwrap();
-        repo.prepare_dir().expect("prepare_dir は成功すべき");
-
-        // 事前: vault.db を保存
-        let header =
-            VaultHeader::new_plaintext(VaultVersion::CURRENT, time::OffsetDateTime::now_utc())
-                .expect("header は常に有効");
-        let initial_vault = Vault::new(header);
-        repo.save(&initial_vault).expect("初期 vault の保存");
-
-        // act
-        let result = repo.load_or_create_plaintext();
-
-        // assert
-        let vault = result.expect("既存 vault.db があれば Ok が返るべき");
-        assert!(vault.records().is_empty(), "初期 vault はレコード 0 件");
-        assert!(
-            matches!(
-                vault.protection_mode(),
-                shikomi_core::ProtectionMode::Plaintext
-            ),
-            "初期 vault は plaintext モードであるべき"
-        );
-        // vault.db が既存なので生成ログは出力されない
-        assert!(
-            !logs_contain("vault not found"),
-            "既存 vault.db がある場合、生成ログは出力されるべきでない"
-        );
-    }
-
-    // --- TC-UT-141: vault.db 不在 — 空 plaintext vault を生成してログ出力 ---
-
-    #[test]
-    #[tracing_test::traced_test]
-    fn tc_ut_141_load_or_create_plaintext_creates_empty_vault_when_absent() {
-        // REQ-DAEMON-028 / Issue #80
-        let dir = TempDir::new().unwrap();
-        let repo = SqliteVaultRepository::from_directory(dir.path()).unwrap();
-        repo.prepare_dir().expect("prepare_dir は成功すべき");
-        // vault.db を意図的に作成しない
-
-        // act
-        let result = repo.load_or_create_plaintext();
-
-        // assert
-        let vault = result.expect("vault.db 不在でも Ok (空 vault 生成) が返るべき");
-        assert!(
-            vault.records().is_empty(),
-            "新規生成した vault はレコード 0 件であるべき"
-        );
-        assert!(
-            matches!(
-                vault.protection_mode(),
-                shikomi_core::ProtectionMode::Plaintext
-            ),
-            "新規生成した vault は plaintext モードであるべき"
-        );
-        assert!(
-            logs_contain("vault not found; created new plaintext vault at"),
-            "生成ログ 'vault not found; created new plaintext vault at' が出力されるべき"
-        );
-        assert!(
-            logs_contain("hint: to enable encryption"),
-            "暗号化ヒントログが出力されるべき"
-        );
-
-        // 横串アサート（冪等性）: vault.db 生成後に再呼出しても Ok が返る
-        let vault2 = repo
-            .load_or_create_plaintext()
-            .expect("2 回目の呼出し: 既存 vault.db を load して Ok が返るべき");
-        assert!(
-            vault2.records().is_empty(),
-            "2 回目の呼出し: 既存 vault のレコード 0 件"
-        );
-    }
-
-    // --- TC-UT-142: 書き込み不可ディレクトリ（Unix 限定）— エラーが返る ---
-
-    #[cfg(unix)]
-    #[test]
-    fn tc_ut_142_load_or_create_plaintext_returns_error_on_readonly_dir() {
-        // REQ-DAEMON-028 / Issue #80
-        // NOTE: spec は Err(PersistenceError::Io(_)) を期待するが、実装上は
-        // load_inner の verify_dir が 0o500 != 0700 を検出して
-        // PersistenceError::InvalidPermission を先に返す。
-        // どちらも「書き込み不可ディレクトリではエラーが返る」意図に沿う。
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = TempDir::new().unwrap();
-        let repo = SqliteVaultRepository::from_directory(dir.path()).unwrap();
-        // vault.db は作成しない
-        // ディレクトリを 0o500（所有者 r-x）に変更して書き込み不可にする
-        std::fs::set_permissions(dir.path(), std::fs::Permissions::from_mode(0o500))
-            .expect("chmod 0o500 に成功すべき");
-
-        // act
-        let result = repo.load_or_create_plaintext();
-
-        // cleanup: TempDir::drop() が削除できるようパーミッション復元
-        let _ = std::fs::set_permissions(dir.path(), std::fs::Permissions::from_mode(0o700));
-
-        // assert: 書き込み不可ディレクトリではエラーが返る
-        assert!(result.is_err(), "0o500 ディレクトリでは Err が返るべき");
-    }
-}
+mod tests;
